Improved handling of unrecognised residues.
If OpenMM fails to match a residue to a template, a dialog box now 
appears allowing the user to try adding hydrogens, exclude this residue 
from simulations, or give up. If the ignore option is chosen, any atoms 
covalently bonded to the ignored residue (and their associated 
hydrogens) will be fixed in space.