Moved all imports from clipper_python c++ library inside functions
PyBind11 throws an error if you attempt to load the same wrapped class 
from two different libraries. This happens during the build process (the 
library is loaded from the build directory for testing, then the 
installed version is loaded while checking the bundle info). Since 
unloading a loaded module from python is somewhat fraught, it's easier 
to ensure that the compiled module is never imported up-front to begin 
with.