Ensure OPENMM_PLUGIN_DIR is set before first import
Since the OpenMM plugins (including the CPU, OpenCL and CUDA) platforms 
are dynamically linked at runtime, this environment variable MUST be set 
before importing importing anything from openmm. Otherwise, the 
best-case scenario is that no platforms are available. The worst-case is 
that it manages to find a *different* version of the plugin libraries 
elsewhere on the library path, causing some very difficult-to-diagnose 
segmentation faults. The necessary code has been moved to the ISOLDE 
bundle's initialize() method, ensuring it's the very first thing called 
on starting ISOLDE.