Avoid use of Atom.idatm_type for finding nonpolar hydrogens
idatm_type is recalcuated from scratch on the first call after any 
change to the model topology, and can get *very* expensive for large 
models. Hydrophobic hydrogens can instead be quickly and cheaply 
identified by looking to see if they're bonded to carbon (the actual 
implementation of this is in Clipper).