Improved performance of softcore nonbonded forces
The three exponent terms (a, b, c) in the
softcore equation are technically adjustable,
but as pointed out by Peter Eastman fixing these
to constant integer values drastically increases the
ability of OpenMM to optimise the derivative calculations.
In practice this reduces the performance overhead from
90-100% to 10-20% - fast enough so softcore potentials
can be turned on by default.