Corrected MD topology for guanosine derivatives
In the original set of ligand AMBER parameterisations done
by Moriarty & Case, the -NH2 group of guanine was
incorrectly assigned as NH3. This commit corrects that for
all current ligands in the components dictionary containing
"guanosine" in their name. The underlying issue is a wider
one affecting various other ligands; correcting this
completely needs a more considered, general approach.