Moved forcefield loading back to main thread
Forcefield loading was previously happening in a separate thread
managed by `ThreadPoolExecutor` to reduce impact on main-thread
performance. For unknown reasons, this seems to *very* rarely lead
to a segmentation fault when loading the pickled force field. The
good news is that the hit to performance is minimal - I added this
threaded approach at a time when all the ligand definitions were
pre-loaded into the main force field making it very slow to
instantiate. Now that ligands are only added as-needed, startup
time is essentially negligible.